Major Improvements
TheoreticalWire pathing is now much better. It uses a general direction (either Vertical or Horizontal) instead of one of the 4 cardinal direction vectors. There is now a 'strictWithWires' boolean option for when you don't want the path to be able to touch the endpoint of an already existing wire going in the opposite direction. Added more helper methods for the recursive pathing such as getPointInLineWith(p1, p2) and many more