Massive UI Changes
Revamped entire UI; this program now uses multiple JavaFX libraries instead of a single Swing JPanel. JavaFX is much more modern and flexible, and I was able to achieve so much in the 3 days that it's been implemented into this project.

Added selection boxes. Dragging on any part of the EditorPanel that isn't a Selectable/Movable entity will make it so a selection box is created. Obviously this will clear the current selection right as the selection box is created, but if you want to select other entities with a selection box while keeping your previous selection, hold control while you make the selection box.

Defined/Coded an XML structure to use for the Save/Load files of Circuits. Currently the only thing that is implemented is loading xml files (you cannot save your Circuits yet), but it seems to be working perfectly. I think that tomorrow or at least within the week I will implement adding Circuits.

Needless to say given that I coded the ability to load projects, the 'Open File' option on he MenuBar is fully coded. It opens up a JavaFX FileChooser to select a File to load, then asks the user if they want to load the file in a new GUI or open it in the current one. Once I implement saving, I will also make it ask if you would like to save your changes, if you choose the latter option.