HUGE Changes
Completely Re-Coded how transmitting works. Added InputNode and OutputNode, both can't be deleted from the ConnectionList. InputNode's depend on OutputNodes and therefore have a field called 'dependencies' which is all the OutputNodes that the InputNode depends on. If the dependencies is greater than 2, then the node is flagged as illogical. ConnectibleEntities also have dependencies, but instead of depending on OutputNodes, they depend on other entities (associated through OutputNodes). ConnectibleEntities can depend on any number of other ConnectibleEntities. There is also a way to check "super dependencies" which recursively goes through the dependencies, the dependencies of those dependencies etc until it finds an independent Entity (an Entity that only has OutputNodes, no InputNodes). If the algorithm detects that it is going in a circle, it returns null. This null return is useful for checking invalid wires, or gates that rely on themselves or gates that form an invalid loop.