MEGA Efficiency changes.
Throw away any patch notes you took from my last commit, they are all garbage. I COMPLETELY forgot that when you clone circuits, the connections for the entities aren't cloned so it has to call connect check on the every entity when a preview circuit is created. This is mega inefficient. If n represents the number of entities in a circuit, each connect check call takes n iterations. So calling connect check on all entities is n squared. At this rate, cloning circuits is not worth it at all for previews and creates a lot of lag.

To fix this, I added a much-needed feature: undoing and re doing. This was VERY daunting to code at first; in this program I coded the wires to split up and merge together in a lot of cases when they are created near each other (this makes connections more consistent, reliable, and easy to notice), so when you add a wire it might get split into multiple by another wire, or it might merge with another wire. Keeping track of all those sub-created/merged/bisected wire additions seemed like the best option but it definitely wasn't. The way I fixed it is that I made undo delete operations much smarter. When you tell the Circuit to keep track of a Wire creation for undoing/redoing (with <circuit instance>.new EntityAddOperation(<wire instance>)), and the respective EntityRemoveOperation(<wire instance> is created through the EntityAddOperation, it doesn't only look for a wire with the same end points as the one in the EntityAddOperation. Since it knows that Wire could have been merged/bisected after the AddOperation happens, it also looks for wires that are 'eaten' by the EntityAddOperation wire. So any wires that are overlapped/enclosed by the AddOperation wire will ALL be deleted, since they are segments of the original AddOperation anyways. Not only does it look for wires that were 'eaten' by the EntityAdd wire, it also checks to see if the EntityAdd wire is 'eaten' by another wire. I don't want to go too in depth, but see Circuit.EntityDeleteOperation.operate for more.

Next up: Adding a BoundingBox whose entity list is continuously updated every time the screen moves a certain distance. It will be called the screen scope and it will make the program more efficient when there are TONS of entities on it. The screen scope will be every entity that is in the scope of [the WHOLE screen, not just the canvas size, expanded by 10% of the screen width