Major changes to power transmission.
If this is hard to read I'm sorry, this is me THROWING my thoughts onto my screen:

- COMPLETELY removed the need for recalculateTransmissions() and recalculatePowerStatuses() in the Circuit class. They will be removed soon.

+ Revamped DependenyTree class and Dependent interface. DependencyTrees work similar to how they used to work, but are now created to temporarily modify the power statuses of an area then they are removed immediately after.
When an entity gets updated (which happens when other entities around it are removed, and when an entity is added that intercepts it), a method called treeUpdate is called on all of the nodes of the entity (or end points if its a wire).

treeUpdate() (in Dependent interface) - automatically resets all the power statuses of everything that it is connected to, recursively. Based on the Output/Input nodes that are hit during the reset method, it creates DependencyTrees based on the nodes that were hit during the recursive resetting, and then calculates the power status of that tree the same way DependencyTrees have done before. After that, it calls pollParent() on the inputNodes of the tree.
So the order of the DependencyTree work goes:
TreeUpdate():
 - reset wires
 - grow trees based on things hit during reset
 - calculate power status of tree, sign nodes we pass by. If we hit a node we already signed we went in a Circle, so entity signed below its signature in the    FlowSignature class has an invalid tree grown there. If this happens, the method stops here.
 - poll the tree (call pollParent on all the InputNodes)
 - during pollParent(), if any output nodes on the polled parent now have a status that disagrees with their set status, call tree update (this big method) on   that out node.

As a result of all of this, recalculating power statuses no longer scales with the number of entities in the Circuit. Also, it has to do practically no work when creating new wires, because only trees below the current tree that you are creating, are ever updated and this is rare as well because the below ones are only updated if there is a power status change.

Another AMAZING result of this is you can now make logic gates depend on logic gates that depend on them (semi self dependent gates is what I call it). This was never possible with my old algorithm. The way that I safeguard against self dependent gates that can actually cause errors (tree update calculation loops) is to have all the above methods (poll(), pollParent(), treeUpdate(), new DependencyTree() ) all take a FlowSignature parameter. When a trees power status is calculated, it "signs" all of its Input and Output nodes onto the flow signature and passes it on to the next methods in the recursion. If a node on a tree ever runs into a FlowSignature that it has already signed (this is checked in DependencyTree.determinePowerStatus), then everything below it in the signatures list is marked as circular and the tree that found this has its canPoll boolean set to false so its poll check that occurs next in the recursion will fail. The circular mark method grows trees at all the signatures and sets their power status to invalid. Then polls them with a signature of error origin. The special thing about signatures with error origin is that when they poll ConnectibleEntities, treeUpdate() wont be called further if the power status that the poll disagrees with is SELF_DEPENDENt