Changes to Power Transmission
+ Added updateTreesInParallel() to update trees at a level order
+ Added updateTreesInSeries() to update trees downwards one at a time

I'm not sure which one I am going to use yet. This is also serving as a backup because I am probably going to change how the polling works. Instead of asking what future trees need to be updated at the moment of calculating the power status, I want to make it so calculating the power status sends a list of Entities that should be polled instead of sending the results of a poll of that moment. This will make it more accurate.